api/requirements.txt
Flask == 1.1.2
Flask-RESTful == 0.3.8
Flask-JWT-Extended ==  3.24.1
flask-mongoengine ==  0.9.5
passlib ==  1.7.2
dnspython == 1.16.0
google-auth == 1.18.0
requests == 2.24.0
Flask-Cors == 1.10.3
Flask-SocketIO == 4.3.1
Flask-PyNgrok == 1.0.3
Flask-Login =  0.5.0